added 'Delete Multiple Users',  'Delete Multiple Spaces',  'Delete Multiple Roles'. Functionality For Super Admin

+++ apps/server/src/api/roles/deleteRoleApi.ts
@@ -1,9 +1,10 @@
-import { getUserAccessRoles } from "@libs/utils/getUserAccessRoles";
+import { getUserAccessRoles } from '@libs/utils/getUserAccessRoles';
 import { RoleCode, RoleScop } from '@prisma/client';
 import prisma from 'apps/server/src/prisma/PrismaClient';
 
 export default async function deleteRoleApi(req, res) {
-  const {  roleId } = req.params;
+  const { roleId } = req.params;
+  const { roleIds } = req.body;
   try {
     const userAccessRoles = getUserAccessRoles(req.user.roles, [
       { scop: RoleScop.SUPER, code: RoleCode.ADMIN },
@@ -11,13 +12,24 @@ export default async function deleteRoleApi(req, res) {
     if (!userAccessRoles.length) {
       return res.sendStatus(403);
     }
-
-    await prisma.role.delete({
-      where: {
-        id: roleId,
-      },
-    });
+    if (roleId) {
+      await prisma.role.delete({
+        where: {
+          id: roleId,
+        },
+      });
+    }
+    if (roleIds?.length) {
+      await prisma.role.deleteMany({
+        where: {
+          id: {
+            in: roleIds,
+          },
+        },
+      });
+    }
     return res.status(200).json({
+      deleted: { roleIds, roleId },
       message: 'Role deleted successfully',
     });
   } catch (error) {
